requirements.txt
amqp==2.2.2
astroid==1.6.1
autopep8==1.3.4
Babel==2.5.3
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cssselect==1.0.3
cssutils==1.0.2
decorator==4.2.1
defusedxml==0.5.0
Django==2.0.3
django-allauth==0.35.0
django-cors-headers==2.2.0
django-countries==5.2
django-debug-toolbar==1.9.1
django-elasticsearch-debug-toolbar==1.2.0
django-filter==1.1.0
django-map-widgets==0.1.9
django-redis==4.9.0
django-rest-auth==0.9.3
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
elasticsearch==6.1.1
elasticsearch-dsl==6.1.0
envparse==0.2.0
factory-boy==2.10.0
Faker==0.8.11
flower==0.9.2
future==0.16.0
geocoder==1.38.0
html2text==2018.1.9
idna==2.6
ipaddress==1.0.19
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
lazy-object-proxy==1.3.1
lxml==4.1.1
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.6
openapi-codec==1.3.2
Pillow==5.0.0
premailer==3.1.1
psycopg2==2.7.3.2
pycodestyle==2.3.1
PyJWT==1.6.0
pylint==1.8.2
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2018.3
ratelim==0.1.6
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.13.2
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
tornado==5.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wrapt==1.10.11
